TAM_libraries/requirements.txt
backports.entry-points-selectable==1.1.0
biopython==1.78
cycler==0.10.0
distlib==0.3.3
et-xmlfile==1.1.0
filelock==3.3.0
joblib==1.1.0
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.3
openpyxl==3.0.9
pandas==1.2.4
Pillow==8.2.0
platformdirs==2.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==1.0.2
scipy==1.6.3
seaborn==0.11.1
selenium==3.141.0
six==1.16.0
TBB==0.2
threadpoolctl==3.1.0
urllib3==1.26.3
virtualenv==20.8.1
xlrd==2.0.1
